Return an error when sending or receiving from a closed socket.
When closing a UDP socket, the internal socket handle will be now be
set to ZED_NET_INVALID_SOCKET. This is then checked in send() and
receive() and returns an error if applicable. The advantage to this
is that it allows us to check the return value of receive() and abort
the thread if it's no longer valid. This saves us from having to track
a separate flag for determining when the terminate the loop.